Added unauthorized warning messages and temporary fix for detection problem on script start after boot.
If the script was started automatically on system boot it didn't
recognize disc activity and put the system to sleep after the
counter expired. Adding a delay seems to fix this behaviour, but
isn't a nice solution. I'll have to find the cause of the problem
and try to fix that.